Refactor and enhance mind map page with new components and API update handling
