+++ src/app/api/mindmaps/[id]/update/update.js
