+++ src/app/components/mindmap/MindMapNode.js
