+++ src/app/components/mindmap/MindMapSidebar.js
