+++ src/app/components/mindmap/MindMapVisualization.js
