+++ src/app/components/mindmap/NodeDetailPanel.js
