+++ src/app/components/mindmap/nodes/index.js
