+++ src/app/config/index.js
@@ -1,0 +1,13 @@
+export const DB_CONFIG = {
+  // MongoDB configuration
+};
+
+export const MINDMAP_CONFIG = {
+  layouts: {
+    dagre: {
+      rankdir: 'LR',
+      nodesep: 80,
+      ranksep: 120,
+    },
+  },
+};
