+++ src/app/constants/index.js
@@ -1,0 +1,10 @@
+export const SYSTEM_PROMPT = `You are an AI tutor...`;
+
+export const API_TOOLS = [
+  {
+    name: 'create_flashcard_set',
+    description: '...',
+    // ...
+  },
+  // ...
+];
