+++ src/app/utils/errorHandler.js
@@ -1,0 +1,4 @@
+export const handleApiError = (error, customMessage) => {
+  console.error(`${customMessage}:`, error);
+  return NextResponse.json({ error: customMessage }, { status: 500 });
+};
